Move code for recovery in RecoveryTool class
